requirements.txt
langgraph==0.2.16
openai>=1.12.0
psycopg2-binary==2.9.9
pgvector==0.3.2
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
beautifulsoup4==4.12.3
requests==2.31.0
alpaca-trade-api==3.1.1
finnhub-python==2.4.18
httpx==0.26.0
tenacity==8.2.3
python-dotenv==1.0.1
structlog==24.1.0
numpy==1.26.4
tiktoken==0.7.0
python-dateutil==2.8.2
pytest==7.4.4
pytest-asyncio==0.23.3
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets>=9.0,<11
jinja2==3.1.3

